+++ NEWS.md
@@ -4,7 +4,7 @@
 -
 
 ### Fixed
--
+- ([#18718](https://github.com/rstudio/rstudio/issues/18718)): Fixed an issue where an R error during session initialization could leave the session hung forever at startup, with RStudio Server reporting that R was taking longer than usual to start.
 
 ### Dependencies
 -